Move: 로그인 관련 패키지 정리 #3
