+++ src/main/java/com/coastee/server/login/auth/AdminAuth.java
@@ -1,4 +1,4 @@
-package com.coastee.server.auth;
+package com.coastee.server.login.auth;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
@@ -8,5 +8,5 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
 @Target(PARAMETER)
 @Retention(RUNTIME)
-public @interface Auth {
+public @interface AdminAuth {
 }
