+++ src/main/java/com/coastee/server/login/auth/AdminOnly.java
@@ -1,12 +1,12 @@
-package com.coastee.server.auth;
+package com.coastee.server.login.auth;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.Target;
 
-import static java.lang.annotation.ElementType.PARAMETER;
+import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
 
-@Target(PARAMETER)
+@Target(METHOD)
 @Retention(RUNTIME)
-public @interface AdminAuth {
+public @interface AdminOnly {
 }
